Basic adaptations for 1.13.

+++ minecrafttrainsimulator/src/main/java/de/robingrether/mcts/Train.java
@@ -136,7 +136,7 @@ public abstract class Train {
 		}
 	}
 	
-	public short getMapId() {
+	public int getMapId() {
 		return controlPanel.getId();
 	}
